Move the FragmentGroup class out of MaskPuzzleMaskFragment
