+++ Assets/Microgames/MaskPuzzle/Scripts/MaskPuzzleFragmentGroup.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 714bb48a3202644089eb19f2063dba25
+timeCreated: 1521818557
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
